[Setup/#1] fix: 엔티티 PK 타입 int → Long 변경 및 User status enum 적용

+++ src/main/java/com/daehanforeigner/capstone/domain/learning_content/LearningContent.java
@@ -16,7 +16,7 @@ public class LearningContent {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "content_id")
-    private int contentId;
+    private Long contentId;
 
     @ManyToOne
     @JoinColumn(name = "category_id")
